WeatherBit is now a functional alternative to OpenWeatherMap. OpenWeatherMap is used by default, with WeatherBit being an optional api query parameter.
Tried to use a generic parent class that relies on underlying specialised functionality for a more abstract approach that separates areas of concern, but by grabbing a singleton instance of the base API class, I may have broken asynchronicity.

I believe it should be capable of working asyncronously as it's only really reading data once it's instantiated, but I'm not certain of this.

TODO: still not yet implemented functionality to get data for any other time than the current.